+++ kitty/options/parse.py
@@ -1167,7 +1167,7 @@ class Parser:
             raise ValueError(f"The value {val} is not a valid choice for tab_bar_style")
         ans["tab_bar_style"] = val
 
-    choices_for_tab_bar_style = frozenset(('fade', 'hidden', 'powerline', 'separator'))
+    choices_for_tab_bar_style = frozenset(('fade', 'hidden', 'powerline', 'separator', 'slant'))
 
     def tab_fade(self, val: str, ans: typing.Dict[str, typing.Any]) -> None:
         ans['tab_fade'] = tab_fade(val)